ovsdb-idl: add function to set next_remote for jsonrpc reconnect.
OVN SouthBound DB in clustered mode,  when one of raft node down and
back online.  All the ovn-controller clients will not migirate back which
cause RAFT DB clients not in balanced state.

This function provides a way to set the IDL jsonrpc session next_remote.
Which can let caller to set the next reconnect remote which can address
the unbalance issue.

Notice that this function is not actually used anywhere in this patch.
This will be used by OVN, though, since OVN is the primary user of
clustered OVSDB.

Signed-off-by: Zhen Wang <zhewang@nvidia.com>